test_progect/medical_ai/requirements.txt
# ── Core ML ──────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
datasets>=2.17.0
accelerate>=0.27.0
tokenizers>=0.15.0

# ── Tabular ───────────────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0

# ── Explainability ────────────────────────────────────
shap>=0.44.0
captum>=0.7.0

# ── RAG / Vector store ────────────────────────────────
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.4.0

# ── API ───────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.26.0

# ── Database / Cache ──────────────────────────────────
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# ── ONNX export & serving ─────────────────────────────
onnx>=1.15.0
onnxruntime>=1.17.0

# ── Experiment tracking ───────────────────────────────
wandb>=0.16.0
mlflow>=2.10.0

# ── Data pipeline ─────────────────────────────────────
pyarrow>=14.0.0
dvc>=3.38.0

# ── Monitoring & observability ────────────────────────
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
structlog>=24.1.0

# ── Utilities ─────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0
tenacity>=8.2.0
